+++ packages/core/src/types/index.ts
@@ -23,4 +23,8 @@ export { ToSendTransaction, TransactionData }
 
 export { InfinitAction, InfinitActionRecord } from './action'
 export { InfinitCache } from './cache'
-export { InfinitCallback } from './callback'
+export { InfinitOffChainAction, InfinitOffChainActionRecord } from './offChainAction'
+
+export { ActionCallback } from './callback/action'
+export { ContractVerifierCallback } from './callback/contractVerifier'
+export { OffChainActionCallback } from './callback/offChainAction'