+++ packages/modules/init-capital/src/actions/deployInitCapital.test.ts
@@ -2,7 +2,6 @@ import { beforeAll, describe, expect, test } from 'vitest'
 
 import { keccak256, toHex, zeroAddress } from 'viem'
 import { privateKeyToAccount } from 'viem/accounts'
-import { arbitrum } from 'viem/chains'
 
 import { InfinitWallet } from '@infinit-xyz/core'
 
@@ -12,27 +11,26 @@ import { DeployInitCapitalAction } from '@actions/deployInitCapital'
 import { hasRole } from '@actions/subactions/tx-builders/utils'
 
 import { readArtifact } from '@/src/utils/artifact'
-import { TestChain, getForkRpcUrl } from '@infinit-xyz/test'
+import { TestChain, TestInfinitWallet } from '@infinit-xyz/test'
 
 describe('deployInitCapitalAction', () => {
   let action: DeployInitCapitalAction
   let client: InfinitWallet
   let client2: InfinitWallet
 
-  // anvil rpc endpoint
-  const rpcEndpoint = getForkRpcUrl(TestChain.arbitrum)
   // anvil tester pk
-  const privateKey = ANVIL_PRIVATE_KEY
   const oneAddress = ARBITRUM_TEST_ADDRESSES.oneAddress
 
   const weth = ARBITRUM_TEST_ADDRESSES.weth
   beforeAll(() => {
-    const account = privateKeyToAccount(privateKey)
-    client = new InfinitWallet(arbitrum, rpcEndpoint, account)
-    client2 = new InfinitWallet(arbitrum, rpcEndpoint, privateKeyToAccount(ANVIL_PRIVATE_KEY_2))
+    const account1 = privateKeyToAccount(ANVIL_PRIVATE_KEY)
+    const account2 = privateKeyToAccount(ANVIL_PRIVATE_KEY_2)
+
+    client = new TestInfinitWallet(TestChain.arbitrum, account1.address)
+    client2 = new TestInfinitWallet(TestChain.arbitrum, account2.address)
   })
 
-  test('deploy all', async () => {
+  test('deploy all', { retry: 3 }, async () => {
     action = new DeployInitCapitalAction({
       params: {
         proxyAdminOwner: oneAddress,
@@ -43,6 +41,8 @@ describe('deployInitCapitalAction', () => {
         maxLiqIncentiveMultiplier: 100n,
         governor: oneAddress,
         guardian: oneAddress,
+        feeAdmin: oneAddress,
+        treasury: oneAddress,
         doubleSlopeIRMConfigs: [
           {
             name: 'StablecoinIRM',
@@ -71,6 +71,7 @@ describe('deployInitCapitalAction', () => {
     })
     const curRegistry = await action.run({}, undefined, undefined)
 
+    expect(curRegistry.feeVault).not.toBe(zeroAddress)
     expect(curRegistry.proxyAdmin).not.toBe(zeroAddress)
     expect(curRegistry.accessControlManager).not.toBe(zeroAddress)
     expect(curRegistry.initOracleProxy).not.toBe(zeroAddress)
@@ -94,6 +95,21 @@ describe('deployInitCapitalAction', () => {
     for (const irm of Object.values(curRegistry.irms!)) {
       expect(irm).not.toBe(zeroAddress)
     }
+    // check feeInfos of feeVault
+    const feeVaultArtifact = await readArtifact('FeeVault')
+    const feeInfos = await client.publicClient.readContract({
+      address: curRegistry.feeVault!,
+      abi: feeVaultArtifact.abi,
+      functionName: 'getFeeInfos',
+      args: [],
+    })
+    expect(feeInfos[0].admin).equal('0x60045e6DE3080D3a6271E635616dBbC20886dfCb')
+    expect(feeInfos[0].treasury).equal('0x60045e6DE3080D3a6271E635616dBbC20886dfCb')
+    expect(feeInfos[0].feeBps).equal(1000n)
+    expect(feeInfos[1].admin).equal(oneAddress)
+    expect(feeInfos[1].treasury).equal(oneAddress)
+    expect(feeInfos[1].feeBps).equal(9000n)
+
     // check AccessControlManager owner
     const accessControlManagerArtifact = await readArtifact('AccessControlManager')
     const accessControlManagerOwner = await client.publicClient.readContract({