+++ vitest.config.ts
@@ -1,7 +1,9 @@
 import path from 'path'
 import tsconfigPaths from 'vite-tsconfig-paths'
 
-import { defineConfig } from 'vitest/config'
+import { configDefaults, defineConfig } from 'vitest/config'
+
+const MAX_TEST_PARALLELISM = 6
 
 export default defineConfig({
   test: {
@@ -9,19 +11,30 @@ export default defineConfig({
     poolOptions: {
       threads: {
         minThreads: 1,
-        maxThreads: 8,
+        maxThreads: MAX_TEST_PARALLELISM,
       },
     },
     globals: true,
     coverage: {
       reporter: ['text', 'json', 'html', 'json-summary'],
       provider: 'v8',
-      exclude: ['**/*.config.*'],
+      exclude: [
+        ...configDefaults.exclude,
+        'packages/test',
+        'scripts',
+        'packages/**/scripts/**',
+        '**/*.config.*',
+        '**/dist/**',
+        'node_modules/**',
+        '**/artifacts/**',
+        '**/cache/**',
+        '**/contracts/**',
+        '**/__mocks__/**',
+      ],
     },
     testTimeout: 25000,
     hookTimeout: 15000,
     globalSetup: [path.resolve(__dirname, 'packages/test/src/globalSetup.ts')],
-    setupFiles: [path.resolve(__dirname, 'packages/test/src/setup.ts')],
   },
   plugins: [tsconfigPaths()], // to resolve imports using Typescipt's path mapping
 })