+++ packages/modules/uniswap-v3/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @infinit-xyz/uniswap-v3
 
+## 0.0.0-alpha.21
+
+### Patch Changes
+
+- [#111](https://github.com/infinit-xyz/infinit-library/pull/111)
+  [`7ed8c28`](https://github.com/infinit-xyz/infinit-library/commit/7ed8c2885e96b8d9b34e3fb0f07dccd6edb85734) Thanks
+  [@mintcnn](https://github.com/mintcnn)! - Bump core version
+
+## 0.0.0-alpha.20
+
+### Patch Changes
+
+- [`4967948`](https://github.com/infinit-xyz/infinit-library/commit/4967948a32e8519387d53e3ec34242de98eded11) Thanks
+  [@melvinmcrn](https://github.com/melvinmcrn)! - bump core version
+
+- Updated dependencies [[`b9624e1`](https://github.com/infinit-xyz/infinit-library/commit/b9624e1797c500b33ffe4695013ee4d07191c737)]:
+  - @infinit-xyz/core@0.1.0-alpha.8
+
+## 0.0.0-alpha.19
+
+### Minor Changes
+
+- [#99](https://github.com/infinit-xyz/infinit-library/pull/99)
+  [`99f7d41`](https://github.com/infinit-xyz/infinit-library/commit/99f7d41f39712ee23742d13f6c772b35e50d8a82) Thanks
+  [@melvinmcrn](https://github.com/melvinmcrn)! - minor: update type from core
+
+### Patch Changes
+
+- Updated dependencies [[`99f7d41`](https://github.com/infinit-xyz/infinit-library/commit/99f7d41f39712ee23742d13f6c772b35e50d8a82),
+  [`99f7d41`](https://github.com/infinit-xyz/infinit-library/commit/99f7d41f39712ee23742d13f6c772b35e50d8a82),
+  [`99f7d41`](https://github.com/infinit-xyz/infinit-library/commit/99f7d41f39712ee23742d13f6c772b35e50d8a82)]:
+  - @infinit-xyz/core@0.1.0-alpha.7
+
+## 0.0.0-alpha.18
+
+### Patch Changes
+
+- [#81](https://github.com/infinit-xyz/infinit-library/pull/81)
+  [`a851684`](https://github.com/infinit-xyz/infinit-library/commit/a85168472b39588461576567805e133573eb4bc0) Thanks
+  [@tapakornl](https://github.com/tapakornl)! - use contracts from node_modules
+
+- [#91](https://github.com/infinit-xyz/infinit-library/pull/91)
+  [`51eb6d4`](https://github.com/infinit-xyz/infinit-library/commit/51eb6d4ddc3906459a4f6df23b9b992100331919) Thanks
+  [@melvinmcrn](https://github.com/melvinmcrn)! - bump viem version from 2.21.16 to 2.21.28
+
+- Updated dependencies [[`128f5d5`](https://github.com/infinit-xyz/infinit-library/commit/128f5d53bb344eb5078436684413ecd12e86a05b),
+  [`51eb6d4`](https://github.com/infinit-xyz/infinit-library/commit/51eb6d4ddc3906459a4f6df23b9b992100331919)]:
+  - @infinit-xyz/core@0.1.0-alpha.5
+
 ## 0.0.0-alpha.17
 
 ### Minor Changes